Move connectivity check code to its own folder
Deprecation warnings are emitted for anyone using the old locations.
They'll still work.

The reason for moving the code out of interface is that connectivity
checking is becoming complicated and will have some additions soon. It
feels easier to manage outside of the already fairly complicated
interface handling code.
