+++ lib/vintage_net/connectivity/tcp_ping.ex
@@ -1,9 +1,15 @@
-defmodule VintageNet.Interface.InternetTester do
+defmodule VintageNet.Connectivity.TCPPing do
   @moduledoc """
-  This module contains functions for testing whether the Internet is available.
+  Test connectivity by making a connection using TCP
 
-  See the InternetConnectivityChecker for a GenServer that checks on regular
-  intervals and updates VintageNet properties as needed.
+  Connectivity with a remote host can be checked by making a TCP connection to
+  it. The connection either works, the connection is refused, or it times out.
+  The first two cases indicate connectivity.
+
+  Normally ICMP is used for testing connectivity, but that requires the new
+  socket API and a Linux kernel with `net.ipv4.ping_group_range` enabled.  This
+  way usually works unless a device is behind a strict firewall, but there's
+  usually at least one IP address/port on the Internet that they allow.
   """
   @ping_timeout 5_000
 
@@ -24,17 +30,17 @@ defmodule VintageNet.Interface.InternetTester do
           :ok | {:error, ping_error_reason()}
   def ping(ifname, {host, port}) do
     with {:ok, src_ip} <- get_interface_address(ifname),
-         # Note: No support for DNS since DNS can't be forced through
-         # an interface. I.e., errors on other interfaces mess up DNS
-         # even if the one of interest is ok.
+         # Note: No support for DNS since DNS can't be forced through an
+         # interface. I.e., errors on other interfaces mess up DNS even if the
+         # one of interest is ok.
          {:ok, dest_ip} <- VintageNet.IP.ip_to_tuple(host),
          {:ok, tcp} <- :gen_tcp.connect(dest_ip, port, [ip: src_ip], @ping_timeout) do
       _ = :gen_tcp.close(tcp)
       :ok
     else
       {:error, :econnrefused} ->
-        # If the remote refuses the connection, then that means that it received it
-        # and we're connected to the internet!
+        # If the remote refuses the connection, then that means that it
+        # received it and we're connected to the internet!
         :ok
 
       {:error, reason} ->
